# Linearize asm register search