--- a/docs/src/api-reference/AsmRegistryPool.rst
+++ b/docs/src/api-reference/AsmRegistryPool.rst
@@ -1,0 +1,8 @@
+
+.. _asmregistrypool-api-reference:
+
+===============
+AsmRegisterPool
+===============
+
+.. autofunction:: Tensile.AsmRegsiterPool::findFreeRange