--- a/tox.ini
+++ b/tox.ini
@@ -65,6 +65,7 @@ commands =
       {toxinidir}/Tensile/TensileCreateLib/ \
       {toxinidir}/Tensile/Tests/unit/test_TensileCreateLibrary.py \
       {toxinidir}/Tensile/Tests/unit/test_KernelFileContext.py \
+      {toxinidir}/Tensile/Tests/unit/test_AsmRegisterPool.py \
       {posargs}
 
 [testenv:docs]
@@ -85,6 +86,7 @@ commands =
       {toxinidir}/Tensile/TensileCreateLib/ \
       {toxinidir}/Tensile/Tests/unit/test_TensileCreateLibrary.py \
       {toxinidir}/Tensile/Tests/unit/test_KernelFileContext.py \
+      {toxinidir}/Tensile/Tests/unit/test_AsmRegisterPool.py \
       {posargs}
 
 [flake8]