**Summary:**

Profiling revealed an $O(n^2)$ implementation that can be converted into a linear $O(n)$ implementation in `AsmRegisterpool::findFreeRange`, which currently consumes 17% of the runtime of TensileCreateLibrary (under the execution conditions below). This PR updates this function to use a two-pointer algorithm that linearizes the time complexity while also avoiding eager evaluation of register availability.

**Outcomes:**

Runtime reduction when benchmarked against the current implementation is:
```
Before change: 32 jobs, gfx90a:
    7744.10s user 641.74s system 1993% cpu 7:00.56 total <--- before
After change: 32 jobs, gfx90a:
    6546.97s user 648.71s system 1908% cpu 6:17.01 total <--- after
```
Which cuts 43.55 seconds (10.4%) off the wall time of the following invocation:
```
time Tensile/bin/TensileCreateLibrary \
              ../rocBLAS/library/src/blas3/Tensile/Logic/asm_full \
              _build_tcl \
              HIP \
              --merge-files \
              --separate-architecture \
              --lazy-library-loading \
              --no-short-file-names \
              --code-object-version=default \
              --cxx-compiler=amdclang++ \
              --jobs=32 \
              --library-format=msgpack \
              --architecture='gfx90a' \
              --verbose=1 | tee results-jobs-$(date -Iseconds).log
```

**Notable changes:**

The internal call to `isRangeAvailable` has been removed from `findFreeRange`, however it is still used elsewhere in the codebase.

**Testing and Environment:**

Standard Tensile testing; Ubuntu 22.01, ROCm 6.3