@nakajee 2024-08-23
By the way, is the original implementation really O(n2)?
I think both before and after are O(n) (O(n*size))