@bstefanuk 2024-08-26
@nakajee Yes, taking another look regarding the time complexity, I agree that the original implementation is O(n*size), however I do still think the new implementation is O(n) since for each iteration we are only conducting constant time operations, as opposed to potentially checking `size` elements.

Either way it speeds up the build, so it's that's a plus.